requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
openai==0.28.0
python-dotenv==1.0.0
pyaudio==0.2.13
